Pass the single GPU model
This commit should be able to run 12B model on
the 1xV100 240GB memory machine on wechat cluster.
However, all the unloaded chunks are instantly move to
CPU, which makes it really slow.

Also, compare to the original design, the current activation
checkpoint is in float32, this makes the maximum model size or
batch size smaller. In my test, the maximum model could still
reach 12B, but the batch size is reduced from 8 to 4.