Change the param back to fp16 and remove some opt
In GPT3_8B model, if we use fp32 param (and autocast),
the chunk moving time will be 14s out of a total 27s,
however, using fp16 param will reduce the moving time
to 6s. The performance loss is too much to neglected and
we have to give up using autocast and turn to apex O2.